Moved the code for getting XOneFi Routers to its own file under hooks. ProviderScreen now dynamically searches for XOneFi Routers, and removed the log statement which was causing errors
